image_server/requirements.txt
flask

git+https://github.com/openai/clip
tqdm
pillow
torch
torchvision
matplotlib
requests
numpy
